hm2/test/api/requirements.txt
 click==6.7 
 Flask==0.12
 itsdangerous==0.24
 Jinja2==2.9.4
 MarkupSafe==0.23
 Werkzeug==0.11.15
 wheel==0.24.0
 flask_sqlalchemy==2.4.4
 psycopg2-binary==2.8.6
 pandas==1.2.0
